refactor: move useKeyboard into hook
